+++ package-test.js
@@ -80,7 +80,9 @@ const EXCEPTIONS = [
   '@ledgerhq/hw-app-eth', // issue with imports
   '@ledgerhq/live-common', // issue with imports
   'bip39', // breaks
-  'ethers'
+  'ethers',
+  'patch-package',
+  'highcharts'
 ];
 const CUSTOM_DIST = {
   ['babel-core']: 'bridge'