+++ src/modules/access-wallet/software/handlers/web3wallet/index.js
@@ -33,13 +33,16 @@ class Web3Wallet extends WalletInterface {
     return this.web3.eth.sendTransaction(tx);
   }
   signMessage(msg) {
-    return new Promise(resolve => {
+    return new Promise((resolve, reject) => {
       this.web3.eth.personal
         .sign(msg, this.getAddressString())
         .then(hex => {
           resolve(getBufferFromHex(hex));
         })
-        .catch(errorHandler);
+        .catch(err => {
+          errorHandler(err);
+          reject(err);
+        });
     });
   }
 }